requirements.txt
# General purpose
fastapi
pandas
pillow
opencv-python
numpy
pipreqs
zmq
PyYAML
scikit-learn
python-dotenv

# Database
psycopg2-binary

# Huggingface
datasets[vision]
transformers
accelerate
evaluate

# torch
torch
torchvision

# AI
albumentations 
pytesseract

########
# General purpose
fastapi
pandas
pillow
opencv-python
numpy
pipreqs
zmq
PyYAML
scikit-learn
pyvideoreader

# AI
albumentations 
pytesseract
onnxruntime
insightface

# Huggingface
# datasets[vision]
# transformers
# accelerate
# evaluate

# torch
# torch
# torchvision
